# Reduce memory overhead of Tracer#trace