--- a/ddtrace.gemspec
+++ b/ddtrace.gemspec
@@ -59,6 +59,9 @@ Gem::Specification.new do |spec|
 
   # locking transitive dependency of webmock
   spec.add_development_dependency 'addressable', '~> 2.4.0'
+  spec.add_development_dependency 'benchmark-ips', '~> 2.8'
+  spec.add_development_dependency 'benchmark-memory', '~> 0.1'
+  spec.add_development_dependency 'memory_profiler', '~> 0.9'
   spec.add_development_dependency 'redcarpet', '~> 3.4' if RUBY_PLATFORM != 'java'
   spec.add_development_dependency 'pry', '~> 0.10.4'
   spec.add_development_dependency 'pry-stack_explorer', '~> 0.4.9.2'