--- a/lib/ddtrace/ext/system.rb
+++ b/lib/ddtrace/ext/system.rb
@@ -1,0 +1,7 @@
+module Datadog
+  module Ext
+    module System
+      PID = 'system.pid'.freeze
+    end
+  end
+end