@delner 2020-05-26
@brettlangdon FWIW, if there are specific rules we're undertaking here for performance (frozen strings, empty hash constant, avoiding `fetch` or blocks) then we should probably add something to our development guide that outlines performance-driven suggestions for style.